% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banks.R
\name{.qt_read_variable_bank}
\alias{.qt_read_variable_bank}
\title{Reader Functions to Load Information}
\usage{
.qt_read_variable_bank(
  config_path_name,
  path,
  config,
  item_type = "variables",
  class_name = c("qt_bank"),
  validate_value_labels = TRUE,
  value_labels = NULL
)
}
\description{
Reader Functions to Load Information
}
