% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banks.R
\name{print-banks}
\alias{print-banks}
\alias{print.qt_qbank}
\alias{print.qt_genbank}
\alias{print.qt_ctrlbank}
\title{Print qretools Bank Objects}
\usage{
\method{print}{qt_qbank}(x, ...)

\method{print}{qt_genbank}(x, ...)

\method{print}{qt_ctrlbank}(x, ...)
}
\arguments{
\item{x}{A bank object (\code{qt_qbank}, \code{qt_genbank}, or \code{qt_ctrlbank})}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the original object
}
\description{
Print a brief summary of bank objects (question, generated variable, or
control parameter banks).
}
