% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print-variables}
\alias{print-variables}
\alias{print.qt_qvar}
\alias{summary.qt_qvar}
\alias{print.qt_ctrlvar}
\alias{print.qt_genvar}
\title{Print a qretools Variable Object}
\usage{
\method{print}{qt_qvar}(x, ...)

\method{summary}{qt_qvar}(x, ...)

\method{print}{qt_ctrlvar}(x, ...)

\method{print}{qt_genvar}(x, ...)
}
\arguments{
\item{x}{A variable object (\code{qt_qvar}, \code{qt_ctrlvar}, or
\code{qt_genvar})}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the original object

Summarize a qretoos Variable Object

Display more information about an individual variable object. The output
shows the information in \code{print()} methods as well as additional information
such as factor labels, full question or description, etc.
}
\description{
Display a brief one- or two-line summary of an individual variable object.
The output shows the variable ID, storage type in brackets, the title on
the first line, and the surveys used on the second line.
}
\examples{
\dontrun{
qbank <- qt_qbank()
print(qbank$variables[["nhd_sat"]])
# nhd_sat [factor] Neighborhood Satisfaction
#   Surveys: bas-2023, bas-2024, bas-2025
}

}
