% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qretools-config.R
\name{print.qt_config}
\alias{print.qt_config}
\title{Print qretools Configuration}
\usage{
\method{print}{qt_config}(x, ...)
}
\arguments{
\item{x}{A qt_config object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the original object
}
\description{
Print a brief summary of a qretools configuration object.
}
