% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{print.qt_mbank}
\alias{print.qt_mbank}
\title{Print a qt_mbank Object}
\usage{
\method{print}{qt_mbank}(x, ...)
}
\arguments{
\item{x}{A \code{qt_mbank} object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the original object
}
\description{
Display a brief summary of a module bank object.
}
