% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{print.qt_module}
\alias{print.qt_module}
\title{Print a qt_module Object}
\usage{
\method{print}{qt_module}(x, ...)
}
\arguments{
\item{x}{A \code{qt_module} object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the original object
}
\description{
Display a brief summary of an individual module object.
}
