% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{print.qt_qre}
\alias{print.qt_qre}
\title{Print a Survey Questionnaire Object}
\usage{
\method{print}{qt_qre}(x, ...)
}
\arguments{
\item{x}{A \code{qt_qre} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Survey Questionnaire Object
}
