% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{print.qt_qreconfig}
\alias{print.qt_qreconfig}
\title{Print Questionnaire Configuration}
\usage{
\method{print}{qt_qreconfig}(x, ...)
}
\arguments{
\item{x}{A qt_qreconfig object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the original object
}
\description{
Print Questionnaire Configuration
}
