% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banks.R
\name{print.qt_vlabs}
\alias{print.qt_vlabs}
\title{Print qretools Value Labels}
\usage{
\method{print}{qt_vlabs}(x, ...)
}
\arguments{
\item{x}{A qt_vlabs object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the original object
}
\description{
Print a brief summary of value labels object.
}
