% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{summary.qt_qre}
\alias{summary.qt_qre}
\title{Summarize a Survey Questionnaire Object}
\usage{
\method{summary}{qt_qre}(object, ...)
}
\arguments{
\item{object}{A \code{qt_qre} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summarize a Survey Questionnaire Object
}
