% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{summary.qt_qreconfig}
\alias{summary.qt_qreconfig}
\title{Summarize Questionnaire Configuration}
\usage{
\method{summary}{qt_qreconfig}(object, ...)
}
\arguments{
\item{object}{A qt_qreconfig object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the original object
}
\description{
Summarize Questionnaire Configuration
}
